JOO-43 feat(naming-convention) : naming 변경
