+++ backend/app/auth/google_handler.go
@@ -6,10 +6,6 @@ import (
 	"github.com/gin-gonic/gin"
 )
 
-type AccessTokenRequest struct {
-	AccessToken string `json:"accessToken"`
-}
-
 func VerifyGoogleAccessToken(c *gin.Context) {
 	var req AccessTokenRequest
 	if err := c.ShouldBindJSON(&req); err != nil {
