+++ backend/app/auth/google_model.go
@@ -1,0 +1,5 @@
+package auth
+
+type AccessTokenRequest struct {
+	AccessToken string `json:"accessToken"`
+}
